[수정] 메인 탭바 컨트롤러 로직 위치 / [수정] 디렉토리 구조
